+++ app/models/randomization_algorithm/algorithms/minimization.rb
@@ -4,7 +4,7 @@ module RandomizationAlgorithm
   module Algorithms
     # Handles list creation and randomization for minimization randomization
     # schemes.
-    class Minimization < Default
+    class Minimization < OneListOrListsBySite
       attr_reader :randomization_scheme
 
       def initialize(randomization_scheme)
@@ -14,58 +14,6 @@ module RandomizationAlgorithm
       # def randomization_error_message
       #   "Treatment Arms may not be set up correctly."
       # end
-
-      def add_missing_lists!(current_user)
-        if stratify_lists_by_site?
-          list_option_ids = []
-          if number_of_lists > 0 && number_of_lists < RandomizationScheme::MAX_LISTS
-            list_option_ids = @randomization_scheme.stratification_factors
-                                                   .find_by(stratifies_by_site: true)
-                                                   .option_hashes.collect { |i| [i] }
-          end
-          list_option_ids.each do |option_hashes|
-            unless find_list_by_option_hashes(option_hashes)
-              extra_options = option_hashes.select { |oh| oh[:extra] }
-              @randomization_scheme.lists.create(
-                project_id: @randomization_scheme.project_id,
-                user_id: current_user.id,
-                extra_options: extra_options
-              )
-            end
-          end
-        else
-          if @randomization_scheme.lists.count == 0 && number_of_lists > 0 && number_of_lists < RandomizationScheme::MAX_LISTS
-            @randomization_scheme.lists.create(project_id: @randomization_scheme.project_id, user_id: current_user.id)
-          end
-        end
-      end
-
-      def number_of_lists
-        if stratify_lists_by_site?
-          @randomization_scheme.project.sites.count
-        else
-          super
-        end
-      end
-
-      def find_list_by_criteria_pairs(criteria_pairs)
-        if stratify_lists_by_site?
-          site_only_criteria_pairs = criteria_pairs.select do |stratification_factor_id, _option_id|
-            @randomization_scheme.stratification_factors.where(id: stratification_factor_id, stratifies_by_site: true).count == 1
-          end
-
-          list = nil
-          @randomization_scheme.lists.each do |l|
-            if l.criteria_match?(site_only_criteria_pairs)
-              list = l
-              break
-            end
-          end
-          list
-        else
-          super
-        end
-      end
 
       def randomize_subject_to_list!(subject, list, current_user, criteria_pairs)
         # Remove any blank randomizations
@@ -80,10 +28,6 @@ module RandomizationAlgorithm
       end
 
       protected
-
-      def stratify_lists_by_site?
-        @randomization_scheme.stratification_factors.where(stratifies_by_site: true).count > 0
-      end
 
       def generate_next_randomization!(list, current_user, criteria_pairs)
         criteria_pairs.collect! { |sfid, oid| [sfid.to_i, oid.to_i] }