+++ app/models/randomization_algorithm/algorithms/permuted_block.rb
@@ -18,7 +18,7 @@ module RandomizationAlgorithm
       def add_missing_lists!(current_user)
         stratifications = []
 
-        if number_of_lists > 0 && number_of_lists < RandomizationScheme::MAX_LISTS
+        if number_of_lists.in?(1..RandomizationScheme::MAX_LISTS - 1)
           stratifications = [[]]
           @randomization_scheme.stratification_factors.each do |stratification_factor|
             stratifications = stratifications.product(stratification_factor.option_hashes)
@@ -27,17 +27,16 @@ module RandomizationAlgorithm
         end
 
         stratifications.each do |option_hashes|
-          unless find_list_by_option_hashes(option_hashes)
-            stratification_factor_option_ids = option_hashes.collect { |oh| oh[:stratification_factor_option_id] }
-            options = @randomization_scheme.stratification_factor_options.where(id: stratification_factor_option_ids)
-            extra_options = option_hashes.select { |oh| oh[:extra] }
-            @randomization_scheme.lists.create(
-              project_id: @randomization_scheme.project_id,
-              user_id: current_user.id,
-              options: options,
-              extra_options: extra_options
-            )
-          end
+          next if find_list_by_option_hashes(option_hashes)
+          stratification_factor_option_ids = option_hashes.collect { |oh| oh[:stratification_factor_option_id] }
+          options = @randomization_scheme.stratification_factor_options.where(id: stratification_factor_option_ids)
+          extra_options = option_hashes.select { |oh| oh[:extra] }
+          @randomization_scheme.lists.create(
+            project: @randomization_scheme.project,
+            user: current_user,
+            options: options,
+            extra_options: extra_options
+          )
         end
       end