+++ db/migrate/20180906174841_add_custom_treatment_name_to_randomizations.rb
@@ -1,0 +1,5 @@
+class AddCustomTreatmentNameToRandomizations < ActiveRecord::Migration[5.2]
+  def change
+    add_column :randomizations, :custom_treatment_name, :string
+  end
+end